+++ src/Shared/DB/TokenInfo.cs
@@ -13,9 +13,9 @@
         public string UrlToken { get; set; } = "";
         public string UrlOwnersWallet { get; set; } = "";
         public string UrlChart { get; set; } = "";
-        public int TellMessageIdIsValid { get; set; } = 0;
-        public int TellMessageIdBotVerified { get; set; } = 0;
-        public int TellMessageIdNotVerified { get; set; } = 0;
+        public long TellMessageIdIsValid { get; set; } = 0;
+        public long TellMessageIdBotVerified { get; set; } = 0;
+        public long TellMessageIdNotVerified { get; set; } = 0;
 
         public bool IsValid { get; set; } = false;
         public bool IsProcessed1 { get; set; } = false;
